Creates Menu Hook with DB-Driven Menu
Here you have two things shown, a hardcoded menu that outputs to Smarty Tags for use in template.
Following Issue#9's db scheme provided by Nands, Create a menu with a Title "UserMenu" or "User_Menu" and add links to it.
A tag will instantly be made available to your templates called "TOP_MENU_USERMENU" or "TOP_MENU_USER_MENU" that will output the menu.
Currently my ForEach loop doesn't take into effect hierachies, so that still needs to be worked on. But you can get a good picture of this.
This is based off of my Menu Hook for the ezRPG Rework at nands/ezrpg.